Add Chat component to App.jsx
